refactor(openapi-validator): исключил валидатор для production
