+++ backend/tests/Feature/Products/StoreCategoryActionTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Tests\Feature\Products;
 
 use App\Contract\Error;
-use App\Infrastructure\Middleware\ValidateOpenApiSchemaMiddleware;
+use App\Infrastructure\OpenApiSchemaValidator\ValidateOpenApiSchema;
 use Carbon\Carbon;
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -128,7 +128,7 @@ final class StoreCategoryActionTest extends TestCase
     #[TestDox('Неправильный запрос')]
     public function testBadRequest(array $body): void
     {
-        $body[ValidateOpenApiSchemaMiddleware::VALIDATE_REQUEST_KEY] = false;
+        $body[ValidateOpenApiSchema::VALIDATE_REQUEST_KEY] = false;
 
         $this
             ->postJson('api/products/category', $body)
